Move test data to new module
