+++ haas/tests/_test_case_data.py
@@ -1,0 +1,31 @@
+from haas.testing import unittest
+from ..suite import TestSuite
+
+
+class TestSuiteSubclass(TestSuite):
+
+    pass
+
+
+class TestCaseSubclass(unittest.TestCase):
+
+    def test_method(self):
+        pass
+
+
+class BadlySubclassedTestCase(unittest.TestCase):
+
+    def __init__(self, wrongly_named):
+        unittest.TestCase.__init__(self, wrongly_named)
+
+    def test_method(self):
+        pass
+
+
+class TestSuiteNotSubclass(object):
+
+    def __init__(self, tests=()):
+        self.tests = tests
+
+    def __iter__(self):
+        return iter(self.tests)
